# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "API do controle de homologação de projetos solares da ConectSol.\n\nAutenticação: POST /api/auth/login (e-mail e senha). Use o accessToken como Bearer. Não há auto-cadastro: as contas são criadas por ADMINISTRADOR ou GESTOR em POST /api/usuarios. Não há logout no servidor — a API é stateless e o cliente descarta os tokens.\n\nMudanças de status são feitas por endpoints de ação (ex.: POST /api/pendencias/{id}/resolver), nunca por PUT: é o que garante a automação entre etapas e a auditoria.",
    "title" : "SolarSync API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/clientes" : {
      "get" : {
        "description" : "`nome` casa nome ou UC Coelba. `statusTriagem=AGUARDANDO_VERIFICACAO` é a fila de quem ninguém checou na Coelba ainda; `semConsultaDebito=true` é a fila da consulta de débito, que precede tanto a resolução da pendência quanto o projeto.",
        "operationId" : "listar_5",
        "parameters" : [ {
          "in" : "query",
          "name" : "filtro",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ClienteFiltro"
          }
        }, {
          "in" : "query",
          "name" : "paginacao",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginaResponseClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Lista clientes paginados, com busca e filtros de fila",
        "tags" : [ "Clientes" ]
      },
      "post" : {
        "operationId" : "criar_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClienteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Cadastra um cliente",
        "tags" : [ "Clientes" ]
      }
    },
...
    "/api/clientes/{id}" : {
      "delete" : {
        "operationId" : "excluir_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "Clientes" ]
      },
      "get" : {
        "operationId" : "buscar_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Clientes" ]
      },
      "put" : {
        "operationId" : "atualizar_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClienteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Clientes" ]
      }
    },
    "/api/clientes/{id}/historico" : {
      "get" : {
        "operationId" : "historico_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo da triagem de pendência do cliente",
        "tags" : [ "Clientes" ]
      }
    },
    "/api/clientes/{id}/reverificar" : {
      "post" : {
        "description" : "Para o novo ciclo (uma ampliação meses depois) ou para desfazer marcação errada. Não apaga o projeto já criado.",
        "operationId" : "reverificar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Devolve o cliente para a fila de verificação de pendência",
        "tags" : [ "Clientes" ]
      }
    },
    "/api/clientes/{id}/sem-pendencia" : {
      "post" : {
        "description" : "Etapa 1 do fluxo pelo caminho \"segue direto\". Cria o Projeto do cliente em RECEBIDO e o coloca na fila de consulta de débito. Se já houver projeto em andamento, nenhum novo é criado.",
        "operationId" : "marcarSemPendencia",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Registra que a Coelba foi checada e o cliente não tem pendência",
        "tags" : [ "Clientes" ]
      }
    },
...
    "/api/debitos/cliente/{clienteId}" : {
      "get" : {
        "description" : "Até dois registros: o débito que trava a pendência e o que trava a homologação. Lista vazia significa que ninguém consultou este cliente ainda — que é diferente de não dever nada.",
        "operationId" : "buscarPorCliente",
        "parameters" : [ {
          "in" : "path",
          "name" : "clienteId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DebitoResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Situações de débito de um cliente",
        "tags" : [ "Débitos" ]
      },
      "put" : {
        "description" : "O tipo diz que etapa esta consulta responde. Com PENDENCIA ativo, resolver a pendência é bloqueado com 409; com HOMOLOGACAO ativo, encaminhar o projeto à Coelba é bloqueado com 409. Reconsultar e achar a mesma situação só atualiza a data, sem gerar linha nova de histórico.",
        "operationId" : "registrarConsulta",
        "parameters" : [ {
          "in" : "path",
          "name" : "clienteId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DebitoRegistrarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DebitoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Registra o resultado da consulta de débito",
        "tags" : [ "Débitos" ]
      }
    },
    "/api/debitos/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR. O caminho normal é registrar QUITADO, que preserva o histórico.",
        "operationId" : "excluir_6",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui o registro de débito",
        "tags" : [ "Débitos" ]
      }
    },
    "/api/debitos/{id}/historico" : {
      "get" : {
        "description" : "É desta linha do tempo que sai a métrica de tempo parado por débito.",
        "operationId" : "historico_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo do débito",
        "tags" : [ "Débitos" ]
      }
    },
...
    "/api/pendencias/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR. O caminho normal para registro errado é cancelar por status, que preserva o histórico.",
        "operationId" : "excluir_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui a pendência",
        "tags" : [ "Pendências" ]
      },
      "get" : {
        "operationId" : "buscar_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PendenciaResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Pendências" ]
      },
      "put" : {
        "description" : "Não altera status: use os endpoints de ação.",
        "operationId" : "atualizar_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PendenciaAtualizarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PendenciaResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edita os dados da pendência",
        "tags" : [ "Pendências" ]
      }
    },
...
    "/api/pendencias/{id}/historico" : {
      "get" : {
        "operationId" : "historico_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo das mudanças de status da pendência",
        "tags" : [ "Pendências" ]
      }
    },
...
    "/api/projetos/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR.",
        "operationId" : "excluir_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui o projeto",
        "tags" : [ "Projetos" ]
      },
      "get" : {
        "operationId" : "buscar_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Projetos" ]
      },
      "put" : {
        "description" : "Não altera status: use os endpoints de ação.",
        "operationId" : "atualizar_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjetoAtualizarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edita os dados do projeto",
        "tags" : [ "Projetos" ]
      }
    },
...
    "/api/projetos/{id}/historico" : {
      "get" : {
        "operationId" : "historico_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo das mudanças de status do projeto",
        "tags" : [ "Projetos" ]
      }
    },
...
    "/api/unificacoes/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR.",
        "operationId" : "excluir_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui a unificação",
        "tags" : [ "Unificações" ]
      },
      "get" : {
        "operationId" : "buscar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnificacaoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Unificações" ]
      },
      "put" : {
        "description" : "Não altera os marcos: use as ações de concluir e desligar.",
        "operationId" : "atualizar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UnificacaoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnificacaoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edita os dados da unificação",
        "tags" : [ "Unificações" ]
      }
    },
...
    "/api/usuarios/{id}" : {
      "delete" : {
        "operationId" : "excluir",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "409" : {
            "description" : "Usuário já referenciado no histórico; desative em vez de excluir"
          }
        },
        "tags" : [ "Usuários" ]
      },
      "get" : {
        "operationId" : "buscar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Usuários" ]
      },
      "put" : {
        "operationId" : "atualizar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UsuarioAtualizarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UsuarioResponse"
                }
              }
            },
            "description" : "GESTOR tentando alterar um ADMINISTRADOR"
          }
        },
        "summary" : "Edita nome, e-mail e papéis; a senha tem endpoint próprio",
        "tags" : [ "Usuários" ]
      }
    },
...
    "/api/vistorias/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR.",
        "operationId" : "excluir_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui a vistoria",
        "tags" : [ "Vistorias" ]
      },
      "get" : {
        "operationId" : "buscar_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VistoriaResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Vistorias" ]
      }
    },
...
    "/api/vistorias/{id}/historico" : {
      "get" : {
        "operationId" : "historico",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo da vistoria, incluindo as idas e vindas de reprova",
        "tags" : [ "Vistorias" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ClienteFiltro" : {
        "type" : "object",
        "properties" : {
          "nome" : {
            "type" : "string"
          },
          "semConsultaDebito" : {
            "type" : "boolean"
          },
          "statusTriagem" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "AGUARDANDO_VERIFICACAO", "COM_PENDENCIA", "SEM_PENDENCIA" ]
            }
          }
        }
      },
      "ClienteRequest" : {
        "type" : "object",
        "properties" : {
          "cidade" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "dataPagamento" : {
            "type" : "string",
            "format" : "date"
          },
          "nome" : {
            "type" : "string",
            "maxLength" : 150,
            "minLength" : 0
          },
          "telefone" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          },
          "ucCoelba" : {
            "type" : "string",
            "maxLength" : 30,
            "minLength" : 0
          },
          "vendedor" : {
            "type" : "string",
            "maxLength" : 150,
            "minLength" : 0
          }
        },
        "required" : [ "nome" ]
      },
      "ClienteResponse" : {
        "type" : "object",
        "properties" : {
          "cidade" : {
            "type" : "string"
          },
          "dataPagamento" : {
            "type" : "string",
            "format" : "date"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nome" : {
            "type" : "string"
          },
          "statusTriagem" : {
            "type" : "string",
            "enum" : [ "AGUARDANDO_VERIFICACAO", "COM_PENDENCIA", "SEM_PENDENCIA" ]
          },
          "telefone" : {
            "type" : "string"
          },
          "ucCoelba" : {
            "type" : "string"
          },
          "vendedor" : {
            "type" : "string"
          }
        }
      },
      "ClienteResumoResponse" : {
        "type" : "object",
        "properties" : {
          "cidade" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nome" : {
            "type" : "string"
          },
          "ucCoelba" : {
            "type" : "string"
          }
        }
      },
...
      "DebitoRegistrarRequest" : {
        "type" : "object",
        "properties" : {
          "consultadoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ATIVO", "QUITADO" ]
          },
          "tipo" : {
            "type" : "string",
            "enum" : [ "PENDENCIA", "HOMOLOGACAO" ]
          }
        },
        "required" : [ "status", "tipo" ]
      },
      "DebitoResponse" : {
        "type" : "object",
        "properties" : {
          "atualizadoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "cliente" : {
            "$ref" : "#/components/schemas/ClienteResumoResponse"
          },
          "consultadoPor" : {
            "$ref" : "#/components/schemas/UsuarioResumoResponse"
          },
          "criadoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "detectadoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "diasParado" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "quitadoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ATIVO", "QUITADO" ]
          },
          "tipo" : {
            "type" : "string",
            "enum" : [ "PENDENCIA", "HOMOLOGACAO" ]
          },
          "ultimaConsultaEm" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "HistoricoStatusResponse" : {
        "type" : "object",
        "properties" : {
          "entidadeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "entidadeTipo" : {
            "type" : "string",
            "enum" : [ "CLIENTE", "PENDENCIA", "DEBITO", "PROJETO", "VISTORIA", "UNIFICACAO" ]
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ocorridoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "statusAnterior" : {
            "type" : "string"
          },
          "statusNovo" : {
            "type" : "string"
          },
          "usuarioId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "PaginaResponseClienteResponse" : {
        "type" : "object",
        "properties" : {
          "conteudo" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ClienteResponse"
            }
          },
          "pagina" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tamanho" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElementos" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPaginas" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ultima" : {
            "type" : "boolean"
          }
        }
      },
...
      "PendenciaAtualizarRequest" : {
        "type" : "object",
        "properties" : {
          "observacao" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "responsavelId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tipo" : {
            "type" : "string",
            "enum" : [ "TROCA_TITULARIDADE", "LIGACAO_NOVA", "EXTENSAO_REDE", "AUMENTO_CARGA", "MUDANCA_PADRAO", "DESMEMBRAMENTO", "REGULARIZACAO_CADASTRAL", "DEBITO_VINCULADO", "ADEQUACAO_TECNICA", "OUTRA" ]
          }
        },
        "required" : [ "tipo" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}